refactor(codegen): extrai class_dispatch + coerce de calls.rs
calls.rs (5076 LOC) convertido em pasta calls/ com submodulos:
- coerce.rs (113 LOC) — Number/String/Boolean/isNaN/isFinite coercoes globais
- class_dispatch.rs (528 LOC) — resolve method/getter/setter/init owner +
  emit_virtual_accessor_dispatch + emit_named_method_call +
  lower_class_method_call_with_recv + emit_method_call +
  emit_virtual_dispatch + AccessorKind + fn_name_has_this_param +
  helpers de subclass/accessor.

mod.rs re-exporta os simbolos pub(crate) das fns que outros modulos
usam (operators.rs, members.rs, expressions/mod.rs).

calls/mod.rs: 5076 -> 4481 linhas. Suite verde (11/11 crates).
