+++ crates/rts-codegen/src/codegen/lower/expressions/calls/mod.rs
@@ -1,7 +1,21 @@
+mod class_dispatch;
+mod coerce;
+
+pub(super) use class_dispatch::{
+    AccessorKind, emit_method_call, emit_named_method_call, emit_virtual_accessor_dispatch,
+    fn_name_has_this_param, lower_class_method_call_with_recv, resolve_getter_owner,
+    resolve_init_owner, resolve_method_owner, resolve_setter_owner,
+};
+
 use anyhow::{Result, anyhow};
 use cranelift_codegen::ir::{AbiParam, InstBuilder, Signature, types as cl};
 use cranelift_module::{Linkage, Module};
 use swc_ecma_ast::{CallExpr, Callee, Expr, MemberProp};
+
+use self::coerce::{
+    lower_coerce_is_finite, lower_coerce_is_nan, lower_coerce_to_boolean, lower_coerce_to_number,
+    lower_coerce_to_string,
+};
 
 use crate::abi::lookup;
 use crate::abi::signature::lower_member;
@@ -1307,300 +1321,6 @@ fn lower_js_global_call(
     }
 }
 
-pub(super) fn resolve_method_owner(ctx: &FnCtx, class: &str, method: &str) -> Option<String> {
-    let mut cur = class.to_string();
-    loop {
-        let meta = ctx.classes.get(&cur)?;
-        if meta.methods.iter().any(|m| m == method) {
-            return Some(cur);
-        }
-        match &meta.super_class {
-            Some(parent) => cur = parent.clone(),
-            None => return None,
-        }
-    }
-}
-
-fn resolve_init_owner(ctx: &FnCtx, class: &str) -> Option<String> {
-    let mut cur = class.to_string();
-    loop {
-        let meta = ctx.classes.get(&cur)?;
-        if meta.has_constructor {
-            return Some(cur);
-        }
-        match &meta.super_class {
-            Some(parent) => cur = parent.clone(),
-            None => return None,
-        }
-    }
-}
-
-fn is_subclass_of(ctx: &FnCtx, child: &str, ancestor: &str) -> bool {
-    let mut cur = child.to_string();
-    loop {
-        if cur == ancestor {
-            return true;
-        }
-        let Some(meta) = ctx.classes.get(&cur) else {
-            return false;
-        };
-        match &meta.super_class {
-            Some(parent) => cur = parent.clone(),
-            None => return false,
-        }
-    }
-}
-
-fn resolve_getter_owner(ctx: &FnCtx, class: &str, prop: &str) -> Option<String> {
-    let mut cur = class.to_string();
-    loop {
-        let meta = ctx.classes.get(&cur)?;
-        if meta.getters.iter().any(|g| g == prop) {
-            return Some(cur);
-        }
-        match &meta.super_class {
-            Some(parent) => cur = parent.clone(),
-            None => return None,
-        }
-    }
-}
-
-pub(super) fn resolve_setter_owner(ctx: &FnCtx, class: &str, prop: &str) -> Option<String> {
-    let mut cur = class.to_string();
-    loop {
-        let meta = ctx.classes.get(&cur)?;
-        if meta.setters.iter().any(|s| s == prop) {
-            return Some(cur);
-        }
-        match &meta.super_class {
-            Some(parent) => cur = parent.clone(),
-            None => return None,
-        }
-    }
-}
-
-#[derive(Copy, Clone, PartialEq, Eq)]
-pub(super) enum AccessorKind {
-    Getter,
-    Setter,
-}
-
-fn accessor_mangled(kind: AccessorKind, owner: &str, prop: &str) -> String {
-    match kind {
-        AccessorKind::Getter => class_getter_name(owner, prop),
-        AccessorKind::Setter => class_setter_name(owner, prop),
-    }
-}
-
-fn class_has_accessor(
-    meta: &crate::codegen::lower::ctx::ClassMeta,
-    kind: AccessorKind,
-    prop: &str,
-) -> bool {
-    match kind {
-        AccessorKind::Getter => meta.getters.iter().any(|g| g == prop),
-        AccessorKind::Setter => meta.setters.iter().any(|s| s == prop),
-    }
-}
-
-fn resolve_accessor_owner(
-    ctx: &FnCtx,
-    kind: AccessorKind,
-    class: &str,
-    prop: &str,
-) -> Option<String> {
-    match kind {
-        AccessorKind::Getter => resolve_getter_owner(ctx, class, prop),
-        AccessorKind::Setter => resolve_setter_owner(ctx, class, prop),
-    }
-}
-
-pub(super) fn emit_virtual_accessor_dispatch(
-    ctx: &mut FnCtx,
-    static_class: &str,
-    static_owner: &str,
-    kind: AccessorKind,
-    prop: &str,
-    recv_i64: cranelift_codegen::ir::Value,
-    arg_values: &[cranelift_codegen::ir::Value],
-) -> Result<TypedVal> {
-    let mut overrides: Vec<(String, String)> = Vec::new();
-    for (cname, _meta) in ctx.classes.iter() {
-        if !is_subclass_of(ctx, cname, static_class) {
-            continue;
-        }
-        if let Some(owner) = resolve_accessor_owner(ctx, kind, cname, prop) {
-            overrides.push((cname.clone(), owner));
-        }
-    }
-    let mut distinct: Vec<String> = Vec::new();
-    for (_c, o) in &overrides {
-        if !distinct.contains(o) {
-            distinct.push(o.clone());
-        }
-    }
-    if !distinct.contains(&static_owner.to_string()) {
-        distinct.insert(0, static_owner.to_string());
-    }
-    if distinct.len() == 1 {
-        return emit_named_method_call(
-            ctx,
-            &accessor_mangled(kind, static_owner, prop),
-            recv_i64,
-            arg_values,
-        );
-    }
-
-    let static_fn_name = accessor_mangled(kind, static_owner, prop);
-    let ret_ty = ctx
-        .user_fns
-        .get(&static_fn_name)
-        .and_then(|abi| abi.ret)
-        .unwrap_or(ValTy::I64);
-
-    let class_handle = emit_class_tag_read(ctx, recv_i64, static_class)?;
-
-    let mut ordered: Vec<(String, String)> = overrides
-        .iter()
-        .filter(|(c, _)| {
-            ctx.classes
-                .get(c)
-                .map(|m| class_has_accessor(m, kind, prop))
-                .unwrap_or(false)
-        })
-        .cloned()
-        .collect();
-    ordered.sort_by_key(|(c, _)| {
-        let mut depth = 0;
-        let mut cur = c.clone();
-        while let Some(meta) = ctx.classes.get(&cur) {
-            match &meta.super_class {
-                Some(p) => {
-                    depth += 1;
-                    cur = p.clone();
-                }
-                None => break,
-            }
-        }
-        std::cmp::Reverse(depth)
-    });
-
-    let merge_block = ctx.builder.create_block();
-    let result_param = ctx
-        .builder
-        .append_block_param(merge_block, ret_ty.cl_type());
-    let str_eq = ctx.get_extern(
-        "__RTS_FN_NS_GC_STRING_EQ",
-        &[cl::I64, cl::I64],
-        Some(cl::I64),
-    )?;
-
-    for (cname, owner) in &ordered {
-        let (cn_ptr, cn_len) = ctx.emit_str_literal(cname.as_bytes())?;
-        let from_static = ctx.get_extern(
-            "__RTS_FN_NS_GC_STRING_FROM_STATIC",
-            &[cl::I64, cl::I64],
-            Some(cl::I64),
-        )?;
-        let inst = ctx.builder.ins().call(from_static, &[cn_ptr, cn_len]);
-        let target_handle = ctx.builder.inst_results(inst)[0];
-        let inst = ctx
-            .builder
-            .ins()
-            .call(str_eq, &[class_handle, target_handle]);
-        let cmp = ctx.builder.inst_results(inst)[0];
-        let zero = ctx.builder.ins().iconst(cl::I64, 0);
-        let is_eq =
-            ctx.builder
-                .ins()
-                .icmp(cranelift_codegen::ir::condcodes::IntCC::NotEqual, cmp, zero);
-
-        let then_block = ctx.builder.create_block();
-        let else_block = ctx.builder.create_block();
-        ctx.builder
-            .ins()
-            .brif(is_eq, then_block, &[], else_block, &[]);
-        ctx.builder.switch_to_block(then_block);
-        ctx.builder.seal_block(then_block);
-        let result = emit_named_method_call(
-            ctx,
-            &accessor_mangled(kind, owner, prop),
-            recv_i64,
-            arg_values,
-        )?;
-        let coerced = match ret_ty {
-            ValTy::I32 => ctx.coerce_to_i32(result).val,
-            ValTy::F64 => to_f64(ctx, result),
-            _ => ctx.coerce_to_i64(result).val,
-        };
-        ctx.builder.ins().jump(merge_block, &[coerced.into()]);
-        ctx.builder.switch_to_block(else_block);
-        ctx.builder.seal_block(else_block);
-    }
-
-    let result = emit_named_method_call(
-        ctx,
-        &accessor_mangled(kind, static_owner, prop),
-        recv_i64,
-        arg_values,
-    )?;
-    let coerced = match ret_ty {
-        ValTy::I32 => ctx.coerce_to_i32(result).val,
-        ValTy::F64 => to_f64(ctx, result),
-        _ => ctx.coerce_to_i64(result).val,
-    };
-    ctx.builder.ins().jump(merge_block, &[coerced.into()]);
-    ctx.builder.switch_to_block(merge_block);
-    ctx.builder.seal_block(merge_block);
-    Ok(TypedVal::new(result_param, ret_ty))
-}
-
-fn emit_named_method_call(
-    ctx: &mut FnCtx,
-    fn_name: &str,
-    recv_i64: cranelift_codegen::ir::Value,
-    arg_values: &[cranelift_codegen::ir::Value],
-) -> Result<TypedVal> {
-    let abi = ctx
-        .user_fns
-        .get(fn_name)
-        .ok_or_else(|| anyhow!("user fn `{fn_name}` nao registrada"))?
-        .clone();
-    let mangled: String = format!("__user_{fn_name}");
-    let fn_id = *ctx
-        .extern_cache
-        .get(mangled.as_str())
-        .ok_or_else(|| anyhow!("mangled `{mangled}` nao registrado"))?;
-    let fref = ctx.fref_for_id(fn_id);
-
-    let mut args = Vec::with_capacity(arg_values.len() + 1);
-    args.push(recv_i64);
-    args.extend_from_slice(arg_values);
-    let inst = ctx.builder.ins().call(fref, &args);
-    let results = ctx.builder.inst_results(inst);
-    if let Some(&v) = results.first() {
-        Ok(TypedVal::new(v, abi.ret.unwrap_or(ValTy::I64)))
-    } else {
-        Ok(TypedVal::new(
-            ctx.builder.ins().iconst(cl::I64, 0),
-            ValTy::I64,
-        ))
-    }
-}
-
-fn collect_method_overrides(ctx: &FnCtx, base: &str, method: &str) -> Vec<(String, String)> {
-    let mut out = Vec::new();
-    for (cname, _meta) in ctx.classes.iter() {
-        if !is_subclass_of(ctx, cname, base) {
-            continue;
-        }
-        if let Some(owner) = resolve_method_owner(ctx, cname, method) {
-            out.push((cname.clone(), owner));
-        }
-    }
-    out
-}
-
 pub(super) fn lower_new(ctx: &mut FnCtx, new_expr: &swc_ecma_ast::NewExpr) -> Result<TypedVal> {
     // (#264) Peel TsAs/Paren etc para suportar `new (Animal as any)(...)`.
     let mut callee_expr: &Expr = new_expr.callee.as_ref();
@@ -2312,206 +2032,6 @@ fn lower_super_method_call(
     emit_method_call(ctx, &owner, &method_name, recv_i64, &arg_values)
 }
 
-pub(super) fn lower_class_method_call_with_recv(
-    ctx: &mut FnCtx,
-    class_name: &str,
-    method_name: &str,
-    recv_i64: cranelift_codegen::ir::Value,
-    call: &CallExpr,
-) -> Result<TypedVal> {
-    validate_visibility(ctx, class_name, method_name)?;
-
-    let static_owner = resolve_method_owner(ctx, class_name, method_name).ok_or_else(|| {
-        anyhow!("metodo `{method_name}` nao encontrado em `{class_name}` ou ancestrais")
-    })?;
-
-    let overrides = collect_method_overrides(ctx, class_name, method_name);
-    let mut distinct_owners = Vec::new();
-    for (_c, o) in &overrides {
-        if !distinct_owners.contains(o) {
-            distinct_owners.push(o.clone());
-        }
-    }
-    if !distinct_owners.contains(&static_owner) {
-        distinct_owners.insert(0, static_owner.clone());
-    }
-
-    let abi_static = ctx
-        .user_fns
-        .get(&format!("__class_{static_owner}_{method_name}"))
-        .ok_or_else(|| anyhow!("metodo estatico `{static_owner}.{method_name}` nao registrado"))?
-        .clone();
-    let expected = abi_static.params.len().saturating_sub(1);
-    if call.args.len() != expected {
-        return Err(anyhow!(
-            "metodo `{static_owner}.{method_name}` espera {} argumento(s), recebeu {}",
-            expected,
-            call.args.len()
-        ));
-    }
-    let mut arg_values = Vec::with_capacity(expected);
-    for (a, expected_ty) in call
-        .args
-        .iter()
-        .zip(abi_static.params.iter().skip(1).copied())
-    {
-        if a.spread.is_some() {
-            return Err(anyhow!("spread em chamada de metodo nao suportado"));
-        }
-        let tv = lower_expr(ctx, &a.expr)?;
-        let value = match expected_ty {
-            ValTy::I32 => ctx.coerce_to_i32(tv).val,
-            ValTy::F64 => to_f64(ctx, tv),
-            _ => ctx.coerce_to_i64(tv).val,
-        };
-        arg_values.push(value);
-    }
-
-    if distinct_owners.len() == 1 {
-        return emit_method_call(ctx, &static_owner, method_name, recv_i64, &arg_values);
-    }
-
-    emit_virtual_dispatch(
-        ctx,
-        class_name,
-        method_name,
-        &static_owner,
-        recv_i64,
-        &arg_values,
-        &overrides,
-    )
-}
-
-fn emit_method_call(
-    ctx: &mut FnCtx,
-    owner: &str,
-    method_name: &str,
-    recv_i64: cranelift_codegen::ir::Value,
-    arg_values: &[cranelift_codegen::ir::Value],
-) -> Result<TypedVal> {
-    let fn_name = format!("__class_{owner}_{method_name}");
-    let abi = ctx
-        .user_fns
-        .get(&fn_name)
-        .ok_or_else(|| anyhow!("metodo `{owner}.{method_name}` nao registrado"))?
-        .clone();
-    let mangled: String = format!("__user_{fn_name}");
-    let fn_id = *ctx
-        .extern_cache
-        .get(mangled.as_str())
-        .ok_or_else(|| anyhow!("metodo mangled `{mangled}` faltando"))?;
-    let fref = ctx.fref_for_id(fn_id);
-
-    let mut args = Vec::with_capacity(arg_values.len() + 1);
-    args.push(recv_i64);
-    args.extend_from_slice(arg_values);
-    let inst = ctx.builder.ins().call(fref, &args);
-    let results = ctx.builder.inst_results(inst);
-    if let Some(&v) = results.first() {
-        Ok(TypedVal::new(v, abi.ret.unwrap_or(ValTy::I64)))
-    } else {
-        Ok(TypedVal::new(
-            ctx.builder.ins().iconst(cl::I64, 0),
-            ValTy::I64,
-        ))
-    }
-}
-
-fn emit_virtual_dispatch(
-    ctx: &mut FnCtx,
-    class_name: &str,
-    method_name: &str,
-    static_owner: &str,
-    recv_i64: cranelift_codegen::ir::Value,
-    arg_values: &[cranelift_codegen::ir::Value],
-    overrides: &[(String, String)],
-) -> Result<TypedVal> {
-    let class_handle = emit_class_tag_read(ctx, recv_i64, class_name)?;
-
-    let ret_ty = ctx
-        .user_fns
-        .get(&format!("__class_{static_owner}_{method_name}"))
-        .and_then(|abi| abi.ret)
-        .unwrap_or(ValTy::I64);
-
-    let mut ordered = overrides.to_vec();
-    ordered.sort_by_key(|(c, _)| {
-        let mut depth = 0;
-        let mut cur = c.clone();
-        while let Some(meta) = ctx.classes.get(&cur) {
-            match &meta.super_class {
-                Some(p) => {
-                    depth += 1;
-                    cur = p.clone();
-                }
-                None => break,
-            }
-        }
-        std::cmp::Reverse(depth)
-    });
-
-    let merge_block = ctx.builder.create_block();
-    let result_param = ctx
-        .builder
-        .append_block_param(merge_block, ret_ty.cl_type());
-    let str_eq = ctx.get_extern(
-        "__RTS_FN_NS_GC_STRING_EQ",
-        &[cl::I64, cl::I64],
-        Some(cl::I64),
-    )?;
-
-    for (cname, owner) in &ordered {
-        let (cn_ptr, cn_len) = ctx.emit_str_literal(cname.as_bytes())?;
-        let from_static = ctx.get_extern(
-            "__RTS_FN_NS_GC_STRING_FROM_STATIC",
-            &[cl::I64, cl::I64],
-            Some(cl::I64),
-        )?;
-        let inst = ctx.builder.ins().call(from_static, &[cn_ptr, cn_len]);
-        let target_handle = ctx.builder.inst_results(inst)[0];
-        let inst = ctx
-            .builder
-            .ins()
-            .call(str_eq, &[class_handle, target_handle]);
-        let cmp = ctx.builder.inst_results(inst)[0];
-        let zero = ctx.builder.ins().iconst(cl::I64, 0);
-        let is_eq =
-            ctx.builder
-                .ins()
-                .icmp(cranelift_codegen::ir::condcodes::IntCC::NotEqual, cmp, zero);
-
-        let then_block = ctx.builder.create_block();
-        let else_block = ctx.builder.create_block();
-        ctx.builder
-            .ins()
-            .brif(is_eq, then_block, &[], else_block, &[]);
-
-        ctx.builder.switch_to_block(then_block);
-        ctx.builder.seal_block(then_block);
-        let result = emit_method_call(ctx, owner, method_name, recv_i64, arg_values)?;
-        let coerced = match ret_ty {
-            ValTy::I32 => ctx.coerce_to_i32(result).val,
-            ValTy::F64 => to_f64(ctx, result),
-            _ => ctx.coerce_to_i64(result).val,
-        };
-        ctx.builder.ins().jump(merge_block, &[coerced.into()]);
-
-        ctx.builder.switch_to_block(else_block);
-        ctx.builder.seal_block(else_block);
-    }
-
-    let result = emit_method_call(ctx, static_owner, method_name, recv_i64, arg_values)?;
-    let coerced = match ret_ty {
-        ValTy::I32 => ctx.coerce_to_i32(result).val,
-        ValTy::F64 => to_f64(ctx, result),
-        _ => ctx.coerce_to_i64(result).val,
-    };
-    ctx.builder.ins().jump(merge_block, &[coerced.into()]);
-
-    ctx.builder.switch_to_block(merge_block);
-    ctx.builder.seal_block(merge_block);
-    Ok(TypedVal::new(result_param, ret_ty))
-}
 
 /// Function global (#359): `new Function(...params, body)` variadic.
 /// Concatena params em CSV e chama __RTS_FN_GL_FUNCTION_NEW(params_str, body).
@@ -4963,116 +4483,4 @@ pub(super) fn emit_namespace_constant(
         ));
     }
     Ok(Some(emit_constant_load(ctx, member)?))
-}
-
-fn lower_coerce_is_nan(ctx: &mut FnCtx, call: &CallExpr) -> Result<TypedVal> {
-    use cranelift_codegen::ir::condcodes::FloatCC;
-    use crate::codegen::lower::ctx::{TypedVal, ValTy};
-    let arg = call.args.first().ok_or_else(|| anyhow!("isNaN requires 1 arg"))?;
-    if arg.spread.is_some() {
-        return Err(anyhow!("isNaN: spread arg not supported"));
-    }
-    let tv = super::lower_expr(ctx, &arg.expr)?;
-    let f = super::operators::to_f64(ctx, tv);
-    let result = ctx.builder.ins().fcmp(FloatCC::Unordered, f, f);
-    Ok(TypedVal::new(result, ValTy::Bool))
-}
-
-fn lower_coerce_is_finite(ctx: &mut FnCtx, call: &CallExpr) -> Result<TypedVal> {
-    use cranelift_codegen::ir::condcodes::FloatCC;
-    use crate::codegen::lower::ctx::{TypedVal, ValTy};
-    let arg = call.args.first().ok_or_else(|| anyhow!("isFinite requires 1 arg"))?;
-    if arg.spread.is_some() {
-        return Err(anyhow!("isFinite: spread arg not supported"));
-    }
-    let tv = super::lower_expr(ctx, &arg.expr)?;
-    let f = super::operators::to_f64(ctx, tv);
-    let abs_f = ctx.builder.ins().fabs(f);
-    let inf = ctx.builder.ins().f64const(f64::INFINITY);
-    let result = ctx.builder.ins().fcmp(FloatCC::LessThan, abs_f, inf);
-    Ok(TypedVal::new(result, ValTy::Bool))
-}
-
-fn lower_coerce_to_number(ctx: &mut FnCtx, call: &CallExpr) -> Result<Option<TypedVal>> {
-    use crate::codegen::lower::ctx::{TypedVal, ValTy};
-    if let Some(arg) = call.args.first() {
-        if arg.spread.is_some() {
-            return Ok(None);
-        }
-        let tv = super::lower_expr(ctx, &arg.expr)?;
-        if matches!(tv.ty, ValTy::Handle) {
-            // Delega para __RTS_FN_GL_NUMBER_FROM_STR(handle) -> f64
-            let from_str = ctx.get_extern("__RTS_FN_GL_NUMBER_FROM_STR", &[cl::I64], Some(cl::F64))?;
-            let inst = ctx.builder.ins().call(from_str, &[tv.val]);
-            let v = ctx.builder.inst_results(inst)[0];
-            return Ok(Some(TypedVal::new(v, ValTy::F64)));
-        }
-        let f = super::operators::to_f64(ctx, tv);
-        return Ok(Some(TypedVal::new(f, ValTy::F64)));
-    }
-    let v = ctx.builder.ins().f64const(0.0);
-    Ok(Some(TypedVal::new(v, ValTy::F64)))
-}
-
-fn lower_coerce_to_string(ctx: &mut FnCtx, call: &CallExpr) -> Result<Option<TypedVal>> {
-    if let Some(arg) = call.args.first() {
-        if arg.spread.is_some() {
-            return Ok(None);
-        }
-        let tv = super::lower_expr(ctx, &arg.expr)?;
-        let h = ctx.coerce_to_handle(tv)?;
-        return Ok(Some(h));
-    }
-    let h = ctx.emit_str_handle(b"")?;
-    Ok(Some(h))
-}
-
-fn lower_coerce_to_boolean(ctx: &mut FnCtx, call: &CallExpr) -> Result<Option<TypedVal>> {
-    use cranelift_codegen::ir::condcodes::{FloatCC, IntCC};
-    use crate::codegen::lower::ctx::{TypedVal, ValTy};
-    if let Some(arg) = call.args.first() {
-        if arg.spread.is_some() {
-            return Ok(None);
-        }
-        let tv = super::lower_expr(ctx, &arg.expr)?;
-        // (#550) String coerce: empty string e' falsy. Handle aponta para
-        // Entry::String — usar gc.string_len(handle) > 0.
-        if matches!(tv.ty, ValTy::Handle) {
-            // Caminho generico: chama gc.string_len(handle) e compara > 0.
-            // Para handles nao-string o len volta 0, o que coincide com
-            // Boolean({}) === true em JS — mas em RTS objetos nao-string-like
-            // sao raros como input direto de Boolean(); aceitamos para empty
-            // string virar false (caso comum).
-            let str_len_fn = ctx.get_extern("__RTS_FN_NS_GC_STRING_LEN", &[cl::I64], Some(cl::I64))?;
-            let inst_l = ctx.builder.ins().call(str_len_fn, &[tv.val]);
-            let len = ctx.builder.inst_results(inst_l)[0];
-            let zero = ctx.builder.ins().iconst(cl::I64, 0);
-            let result = ctx.builder.ins().icmp(IntCC::NotEqual, len, zero);
-            return Ok(Some(TypedVal::new(result, ValTy::Bool)));
-        }
-        if matches!(tv.ty, ValTy::F64) {
-            // (#208 fix): NaN deve ser falsy. fcmp NotEqual retorna false
-            // pra NaN (NaN != X e' Unordered, nao true). Use OrderedNotEqual
-            // que e' false quando NaN — exatamente o comportamento desejado.
-            // Em Cranelift: OrderedNotEqual = "ordered AND not equal".
-            let zero = ctx.builder.ins().f64const(0.0);
-            let ne_zero = ctx.builder.ins().fcmp(FloatCC::OrderedNotEqual, tv.val, zero);
-            return Ok(Some(TypedVal::new(ne_zero, ValTy::Bool)));
-        }
-        let v = ctx.coerce_to_i64(tv).val;
-        let zero = ctx.builder.ins().iconst(cl::I64, 0);
-        let result = ctx.builder.ins().icmp(IntCC::NotEqual, v, zero);
-        return Ok(Some(TypedVal::new(result, ValTy::Bool)));
-    }
-    let v = ctx.builder.ins().iconst(cl::I64, 0);
-    Ok(Some(TypedVal::new(v, ValTy::Bool)))
-}
-
-/// Retorna true quando a função compilada tem `this` como primeiro parâmetro.
-/// Isso ocorre em métodos de classe não-estáticos: nome começa com `__class_`
-/// e não contém `_static_` e não é o init synthetic (`__init`).
-pub(super) fn fn_name_has_this_param(name: &str) -> bool {
-    name.starts_with("__class_")
-        && !name.contains("_static_")
-        && !name.ends_with("__init")
 }
